Split out encoding functions.
